bi_engine/requirements.txt
asyncpg>=0.29.0
psycopg2-binary
requests>=2.31.0
cleanco>=2.1
pandas
jellyfish>=1.0.3
python-dotenv>=1.0.1
playwright
beautifulsoup4
pytesseract
opencv-python-headless
Pillow
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.7.0
apscheduler
python-telegram-bot
resend
anthropic>=0.25.0
cachetools>=5.3.3
pytest>=8.1.0
pytest-asyncio>=0.23.6
